Fixed build errors for gopherjs
