+++ assets_desktop.go
@@ -12,39 +12,6 @@ type Image interface {
 	Height() int
 }
 
-func LoadShader(vertSrc, fragSrc string) *gl.Program {
-	vertShader := Gl.CreateShader(Gl.VERTEX_SHADER)
-	Gl.ShaderSource(vertShader, vertSrc)
-	Gl.CompileShader(vertShader)
-	defer Gl.DeleteShader(vertShader)
-
-	fragShader := Gl.CreateShader(Gl.FRAGMENT_SHADER)
-	Gl.ShaderSource(fragShader, fragSrc)
-	Gl.CompileShader(fragShader)
-	defer Gl.DeleteShader(fragShader)
-
-	program := Gl.CreateProgram()
-	Gl.AttachShader(program, vertShader)
-	Gl.AttachShader(program, fragShader)
-	Gl.LinkProgram(program)
-
-	return program
-}
-
-func (r *Region) Texture() *gl.Texture {
-	return r.texture.id
-}
-
-type Texture struct {
-	id     *gl.Texture
-	width  float32
-	height float32
-}
-
-func (t *Texture) Texture() *gl.Texture {
-	return t.id
-}
-
 func NewTexture(img Image) *Texture {
 	var id *gl.Texture
 	if !headless {
