+++ assets_js.go
@@ -3,6 +3,7 @@
 package engo
 
 import (
+	"engo.io/gl"
 	"github.com/gopherjs/gopherjs/js"
 )
 
@@ -12,41 +13,8 @@ type Image interface {
 	Height() int
 }
 
-func LoadShader(vertSrc, fragSrc string) *js.Object {
-	vertShader := Gl.CreateShader(Gl.VERTEX_SHADER)
-	Gl.ShaderSource(vertShader, vertSrc)
-	Gl.CompileShader(vertShader)
-	defer Gl.DeleteShader(vertShader)
-
-	fragShader := Gl.CreateShader(Gl.FRAGMENT_SHADER)
-	Gl.ShaderSource(fragShader, fragSrc)
-	Gl.CompileShader(fragShader)
-	defer Gl.DeleteShader(fragShader)
-
-	program := Gl.CreateProgram()
-	Gl.AttachShader(program, vertShader)
-	Gl.AttachShader(program, fragShader)
-	Gl.LinkProgram(program)
-
-	return program
-}
-
-func (r *Region) Texture() *js.Object {
-	return r.texture.id
-}
-
-type Texture struct {
-	id     *js.Object
-	width  float32
-	height float32
-}
-
-func (t *Texture) Texture() *js.Object {
-	return t.id
-}
-
 func NewTexture(img Image) *Texture {
-	var id *js.Object
+	var id *gl.Texture
 	if !headless {
 		id = Gl.CreateTexture()
 		Gl.BindTexture(Gl.TEXTURE_2D, id)
@@ -54,10 +22,11 @@ func NewTexture(img Image) *Texture {
 		Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_MAG_FILTER, Gl.LINEAR)
 		Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_MIN_FILTER, Gl.LINEAR_MIPMAP_NEAREST)
 		Gl.GenerateMipmap(Gl.TEXTURE_2D)
-		Gl.BindTexture(Gl.TEXTURE_2D, nil)
-		Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_WRAP_S, Gl.CLAMP_TO_EDGE)
-		Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_WRAP_T, Gl.CLAMP_TO_EDGE)
-
+		/*
+			Gl.BindTexture(Gl.TEXTURE_2D, nil)
+			Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_WRAP_S, Gl.CLAMP_TO_EDGE)
+			Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_WRAP_T, Gl.CLAMP_TO_EDGE)
+		*/
 		if img.Data() == nil {
 			panic("Texture image data is nil.")
 		}
